@@ -1,9 +1,10 @@
 buildscript {
     repositories {
+        google()
         jcenter()
     }
     dependencies {
-        classpath 'com.android.tools.build:gradle:2.3.3'
+        classpath 'com.android.tools.build:gradle:3.6.3'
     }
 }
 
@@ -17,19 +18,18 @@ allprojects {
 }
 
 ext {
-    minSdkVersion = 16
-    sampleMinSdkVersion = 16
-    targetSdkVersion = 25
-    compileSdkVersion = 25
-    buildToolsVersion = "25.0.3"
+    minSdkVersion = 21
+    sampleMinSdkVersion = 21
+    targetSdkVersion = 29
+    compileSdkVersion = 29
+    buildToolsVersion = "29.0.3"
 }
 
 ext.deps = [
         // Android
-        recyclerview         : 'com.android.support:recyclerview-v7:25.4.0',
-        supportv4            : 'com.android.support:support-v4:25.4.0',
-        appcompat            : 'com.android.support:appcompat-v7:25.4.0',
-        cardview             : 'com.android.support:cardview-v7:25.4.0',
+        recyclerview         : 'androidx.recyclerview:recyclerview:1.1.0',
+        appcompat            : 'androidx.appcompat:appcompat:1.1.0',
+        cardview             : 'androidx.cardview:cardview:1.0.0',
 
         // External
         recyclerviewanimators: 'jp.wasabeef:recyclerview-animators:2.2.7',
